@@ -39,7 +40,7 @@
     "file-saver": "^2.0.5",
     "filesize": "^10.0.6",
     "highlight.js": "^11.7.0",
-    "next": "^13.1.7-canary.7",
+    "next": "13.2.5-canary.17",
     "next-seo": "^5.15.0",
     "next-themes": "^0.2.1",
     "pdfjs-dist": "^3.3.122",
