@@ -10,39 +10,39 @@ const WordCounter = () => {
   const [value, setValue] = React.useState('')
   const words = value.match(/\S+/g)?.length || 0
   const chars = value.length || 0
-  const charsWithoutSpace = value.replace(/ /g, '').length || 0
+  const charsWithoutSpaces = value.replace(/ /g, '').length || 0
   const paragraphs =
     value.split('\n').filter((paragraph) => paragraph !== '').length || 0
 
   return (
-    <Container className='max-w-5xl flex justify-center flex-col items-center'>
-      <Title title='字數計算器' />
+    <Container className='flex max-w-5xl flex-col items-center justify-center'>
+      <Title title='Word Counter' />
 
-      {/* 字數計算器 */}
-      <div className='w-full grid my-12 grid-cols-2 sm:grid-cols-4 gap-2'>
-        <div className='border border-accent-2 p-3 h-24 rounded-lg w-full'>
-          <div className='font-bold text-2xl'>{words}</div>
-          <div className='text-accent-6 text-xs font-bold'>個字</div>
+      {/* Word counter */}
+      <div className='my-12 grid w-full grid-cols-2 gap-2 sm:grid-cols-4'>
+        <div className='h-24 w-full rounded-lg border border-accent-2 p-3'>
+          <div className='text-2xl font-bold'>{words}</div>
+          <div className='text-xs font-bold text-accent-6'>words</div>
         </div>
-        <div className='border border-accent-2 p-3 h-24 rounded-lg w-full'>
-          <div className='font-bold text-2xl'>{chars}</div>
-          <div className='text-accent-6 text-xs font-bold'>個字符</div>
+        <div className='h-24 w-full rounded-lg border border-accent-2 p-3'>
+          <div className='text-2xl font-bold'>{chars}</div>
+          <div className='text-xs font-bold text-accent-6'>characters</div>
         </div>
-        <div className='border border-accent-2 p-3 h-24 rounded-lg w-full'>
-          <div className='font-bold text-2xl'>{charsWithoutSpace}</div>
-          <div className='text-accent-6 text-xs font-bold'>
-            個空格以外的字符
+        <div className='h-24 w-full rounded-lg border border-accent-2 p-3'>
+          <div className='text-2xl font-bold'>{charsWithoutSpaces}</div>
+          <div className='text-xs font-bold text-accent-6'>
+            characters without spaces
           </div>
         </div>
-        <div className='border border-accent-2 p-3 h-24 rounded-lg w-full'>
-          <div className='font-bold text-2xl'>{paragraphs}</div>
-          <div className='text-accent-6 text-xs font-bold'>段</div>
+        <div className='h-24 w-full rounded-lg border border-accent-2 p-3'>
+          <div className='text-2xl font-bold'>{paragraphs}</div>
+          <div className='text-xs font-bold text-accent-6'>paragraphs</div>
         </div>
       </div>
 
-      {/* 文字區域 */}
+      {/* Textarea */}
       <TextareaAutosize
-        className='rounded-md border border-accent-2 bg-hong-bg py-2 px-3 transition-colors duration-200 ease-linear focus:border-accent-5 focus:outline-none w-full'
+        className='w-full rounded-md border border-accent-2 bg-hong-bg py-2 px-3 transition-colors duration-200 ease-linear focus:border-accent-5 focus:outline-none'
         placeholder='在這裡輸入 ...'
         value={value}
         onChange={(e) => setValue(e.target.value)}
