@@ -13,7 +13,7 @@ const Select = (props: SelectProps) => {
         valueContainer: () => 'px-2 py-0.5',
         indicatorsContainer: () => '[&>div]:p-2',
         indicatorSeparator: () => 'w-px bg-accent-2 my-2 self-stretch',
-        option: () => 'p-2 hover:bg-theme-4 active:bg-theme-5',
+        option: () => 'p-2 hover:bg-accent-2',
         menu: () =>
           'border border-accent-2 rounded-lg absolute left-0 top-[calc(100%+12px)] bg-hong-bg',
         ...classNames,
